python/requirements.txt
fastapi
uvicorn
pydantic
numpy
opencv-python
ultralytics
ffmpeg-python
pydub
speechrecognition
